+++ source/extensions/lightspeed.trex.viewports.manipulators/docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 ﻿# Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.1.0] - 2023-12-11
+### Added
+- Global selection manager
+
 ## [1.0.0] - 2023-05-18
 ### Added
 - Created