+++ source/extensions/lightspeed.trex.viewports.shared.widget/docs/CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.1.0] - 2023-12-11
+### Changed
+- Expose viewport layer instances
+- Remove legacy light gizmo
+
 ## [1.0.1] - 2023-11-29
 ### Changed
 - Shared viewport will be deactivated until clicked on